app/apps/requirements.txt
streamlit==1.37.1
pandas==2.2.2
matplotlib==3.9.0
gspread==6.1.2
google-auth==2.31.0
gspread-dataframe==3.3.1
